Move 3rdparty libs dir out from UltraCanvas sources
Fix vips for Windows
chanaged Windows build from vcpkg to MingW

+++ UltraCanvas/OS/Linux/UltraCanvasLinuxWindow.cpp
@@ -633,7 +633,7 @@ namespace UltraCanvas {
     }
 
 // ===== ACCESSORS =====
-    unsigned long UltraCanvasLinuxWindow::GetNativeHandle() const {
+    uintptr_t UltraCanvasLinuxWindow::GetNativeHandle() const {
         return xWindow;
     }
